Wrap jQuery plugin in a self-executing anonymous function.
By wrapping a jQuery plugin in a self-executing anonymous function, and passing in jQuery as an argument, the plugin author can SAFELY use the $ alias within the plugin code and still know that the code will work whether or not jQuery.noConflict() is called by someone who uses the plugin.  This also prevents any variables that are introduced during the plugin's creation from leaking out into the global scope.