Added options "headings" and "class"
Congratulations for this plugin and thanks!

This PR would like to contribute by adding the following options:

 * headings
 * class

Example [here](https://ircama.github.io/osm-carto-tutorials/tile-server-ubuntu/) (unfinished document at the time of writing).

I also updated index.md (the documentation) and I set the version to "0.4.2".

Feel free to revise the code as you like.

Notice that with this PR tests and CDN are not updated.

 Changes to be committed:
	modified:   bootstrap-toc.js
	modified:   dist/bootstrap-toc.js
	modified:   dist/bootstrap-toc.min.js
	modified:   index.md
	modified:   package.json